final typesetting fix

+++ koopmans/calculators/_ui/_calculator.py
@@ -350,9 +350,9 @@ class UnfoldAndInterpolateCalculator(CalculatorExt, Calculator, CalculatorABC):
         for kpt in self.parameters.kpath.kpts:
             kvec.append(crys_to_cart(kpt, self.atoms.acell.reciprocal(), +1))
 
-        kx: List[np.float_] = [0.]
+        kx: List[float] = [0.0]
         for ik in range(1, len(kvec)):
-            dxmod = np.linalg.norm(kvec[ik] - kvec[ik - 1])
+            dxmod = float(np.linalg.norm(kvec[ik] - kvec[ik - 1]))
             if ik == 1:
                 dxmod_save = dxmod
             if dxmod > 5 * dxmod_save:
